requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
gevent==23.9.1
gevent-websocket==0.10.1
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.12.0
email-validator==2.0.0.post2
WTForms==3.0.1
psutil==5.9.8 